+++ lib/python/testplan/__init__.py
@@ -1,0 +1,1 @@
+"""Reusable pytest/xmosdoc test plan helpers."""